requirements.txt
pandas
schedule
python-dotenv
requests
beautifulsoup4
selenium